# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "LIS",
        "description": "Laboratory Information Service API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8089",
            "description": "Local server"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
    "paths": {
...
        "/api/v1/lis-service/facilities": {
            "get": {
                "tags": [
                    "Facilities"
                ],
                "summary": "List facilities (paged), optionally filtered by district and active flag",
                "operationId": "list_7",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 20
                        }
                    },
                    {
                        "name": "districtId",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "isActive",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PagedResponseFacilityResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/lis-service/facilities/{id}": {
            "get": {
                "tags": [
                    "Facilities"
                ],
                "summary": "Get a facility by id",
                "operationId": "getById_7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/FacilityResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/lis-service/districts": {
            "get": {
                "tags": [
                    "Districts"
                ],
                "summary": "List districts (paged), optionally filtered by province",
                "operationId": "list_8",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 20
                        }
                    },
                    {
                        "name": "provinceId",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PagedResponseDistrictResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/lis-service/districts/{id}": {
            "get": {
                "tags": [
                    "Districts"
                ],
                "summary": "Get a district by id",
                "operationId": "getById_8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/DistrictResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "FacilityResponse": {
                "type": "object",
                "description": "Health facility",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "description": "Facility id",
                        "example": 1001
                    },
                    "name": {
                        "type": "string",
                        "description": "Facility name",
                        "example": "University Teaching Hospital"
                    },
                    "districtId": {
                        "type": "integer",
                        "format": "int64",
                        "description": "District id",
                        "example": 42
                    },
                    "hmisCode": {
                        "type": "string",
                        "description": "HMIS code",
                        "example": "504010"
                    },
                    "mflCode": {
                        "type": "string",
                        "description": "Master Facility List code",
                        "example": "1010101"
                    },
                    "isActive": {
                        "type": "boolean",
                        "description": "Whether the facility is active",
                        "example": true
                    }
                }
            },
            "PagedResponseFacilityResponse": {
                "type": "object",
                "description": "Paged response",
                "properties": {
                    "content": {
                        "type": "array",
                        "description": "List of items",
                        "example": "[{\"id\":1,\"name\":\"Item 1\"}, ...]",
                        "items": {
                            "$ref": "#/components/schemas/FacilityResponse"
                        }
                    },
                    "totalElements": {
                        "type": "integer",
                        "format": "int64",
                        "description": "Total number of items",
                        "example": 100
                    },
                    "totalPages": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Total number of pages",
                        "example": 10
                    },
                    "currentPage": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Current page number",
                        "example": 1
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Number of items per page",
                        "example": 20
                    },
                    "hasNext": {
                        "type": "boolean"
                    },
                    "hasPrevious": {
                        "type": "boolean"
                    }
                }
            },
            "DistrictResponse": {
                "type": "object",
                "description": "District",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "description": "District id",
                        "example": 42
                    },
                    "name": {
                        "type": "string",
                        "description": "District name",
                        "example": "Kafue"
                    },
                    "provinceId": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Owning province id",
                        "example": 1
                    },
                    "code": {
                        "type": "string",
                        "description": "District code",
                        "example": "KAF"
                    }
                }
            },
            "PagedResponseDistrictResponse": {
                "type": "object",
                "description": "Paged response",
                "properties": {
                    "content": {
                        "type": "array",
                        "description": "List of items",
                        "example": "[{\"id\":1,\"name\":\"Item 1\"}, ...]",
                        "items": {
                            "$ref": "#/components/schemas/DistrictResponse"
                        }
                    },
                    "totalElements": {
                        "type": "integer",
                        "format": "int64",
                        "description": "Total number of items",
                        "example": 100
                    },
                    "totalPages": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Total number of pages",
                        "example": 10
                    },
                    "currentPage": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Current page number",
                        "example": 1
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Number of items per page",
                        "example": 20
                    },
                    "hasNext": {
                        "type": "boolean"
                    },
                    "hasPrevious": {
                        "type": "boolean"
                    }
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}